Fit a geeglm model using AIPW

provides augmented inverse probability weighted estimates of parameters for GEE model of response variable using different covariance structure

AIPW(
  data,
  formula,
  id,
  visit,
  family,
  init.beta = NULL,
  init.alpha = NULL,
  init.phi = NULL,
  tol = 0.001,
  weights = NULL,
  corstr = "independent",
  maxit = 50,
  m = 2,
  pMat,
  method = NULL
)

Arguments

data

longitudinal data set where each subject's outcome has been measured at same time points and number of visits for each patient is similar. Covariance structure of the outcome variable like "unstructured","independent","AR-1" ,"exchangeable"

formula

formula for the response model

id

column name of id of subjects in the dataset

visit

column name of timepoints of visit in the dataset

family

name of the distribution for the response variable, For more information on how to use family objects, see family

init.beta

initial values for the regression coefficient of GEE model

init.alpha

initial values for the correlation structure

init.phi

initial values for the csale parameter for

tol

tolerance in calculation of coefficients

weights

A vector of weights for each observation. If an observation has weight 0, it is excluded from the calculations of any parameters. Observations with a NA anywhere (even in variables not included in the model) will be assigned a weight of 0. Weights are updated as the mentioned the details.

corstr

a character string specifying the correlation structure. It could "independence", "exchangeable", "AR-1", "unstructured"

maxit

maximum number iteration for newton-raphson

m

number of imputation used to update the missing score function value due incomplete data.

pMat

predictor matrix as obtained in mice

method

method option for mice model,for information see mice

Value

A list of objects containing the following objects

call

details about arguments passed in the function

beta

estimated regression coeffictient value for the response model

niter

number of iteration required

betalist

list of beta values at different iteration

weight

estimated weights for the observations

mu

mu values according glm

phi

etsimated phi value for the glm model

hessian

estimated hessian matrix obtained from the last iteration

betaSand

sandwich estimator value for the variance covariance matrix of the beta

Details

AIPW

It uses the inverse probability weighted method to reduce the bias due to missing values in GEE model for longitudinal data. The response variable \(\mathbf{Y}\) is related to the coariates as \(g(\mu)=\mathbf{X}\beta\), where g is the link function for the glm. The estimating equation is $$\sum_{i=1}^{k}\sum_{j=1}^{n}(\frac{\delta_{ij}}{\pi_{ij}}S(Y_{ij},\mathbf{X}_{ij},\mathbf{X}'_{ij})+(1-\frac{\delta_{ij}}{\pi_{ij}})\phi(\mathbf{V}=\mathbf{v}))=0$$ where \(\delta_{ij}=1\) if there is missing value in covariates and 0 otherwise, \(\mathbf{X}\) is fully observed all subjects and \(\mathbf{X}'\) is partially missing, where \(\mathbf{V}=(Y,\mathbf{X})\). The missing score function values due to incomplete data are estimated using an imputation model through mice which we have considered as \(\phi(\mathbf{V}=\mathbf{v}))\).

References

Wang, C. Y., Shen-Ming Lee, and Edward C. Chao. "Numerical equivalence of imputing scores and weighted estimators in regression analysis with missing covariates." Biostatistics 8.2 (2007): 468-473.

Seaman, Shaun R., and Stijn Vansteelandt. "Introduction to double robust methods for incomplete data." Statistical science: a review journal of the Institute of Mathematical Statistics 33.2 (2018): 184.

Vansteelandt, Stijn, James Carpenter, and Michael G. Kenward. "Analysis of incomplete data using inverse probability weighting and doubly robust estimators." Methodology: European Journal of Research Methods for the Behavioral and Social Sciences 6.1 (2010): 37.

See also

Author

Atanu Bhattacharjee, Bhrigu Kumar Rajbongshi and Gajendra Kumar Vishwakarma

Examples

 if (FALSE) {
##
formula<-C6kine~ActivinRIB+ActivinRIIA+ActivinRIIAB+Adiponectin+AgRP+ALCAM
pMat<-mice::make.predictorMatrix(srdata1[names(srdata1)%in%all.vars(formula)])
m1<-AIPW(data=srdata1,
formula<-formula,id='ID',
visit='Visit',family='gaussian',init.beta = NULL,
init.alpha=NULL,init.phi=1,tol=.00001,weights = NULL,
corstr = 'exchangeable',maxit=50,m=3,pMat=pMat)
##
}